+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG for Sulu Pricing Bundle
 =================================
 
+* dev-develop
+
+    * BUGFIX  Fixed calculation of addon items
+    * FEATURE Added config for defining default currency for price calculation
+
 * 0.4.0 (2016-07-14)
 
     * FEATURE Added handling of recurring-prices